fix: fixed the BytesToGB function

+++ util/conversion.go
@@ -1,6 +1,6 @@
 package util
 
 func BytesToGB(bytes int) float64 {
-	gigabytes := int(bytes) / (1e+9)
+	gigabytes := float64(bytes) / (1e+9)
 	return gigabytes
 }
